Seperate Stage 2 into different components.
